requirements.txt
flake8
jinja2
python-barcode
